Usability enhancements

+++ volume_encryption.py
@@ -5,42 +5,71 @@ Overview:
     Iterate through each attached volume and encrypt it for EC2.
 Params:
     ID for EC2 instance
-    Customer Master Key (CMK) (optional)
-    Profile to use
+    KMS key ID
 Conditions:
-    Return if volume already encrypted
-    Use named profiles from credentials file
+    Return if volumes are already encrypted with the provided key
 """
 
+import argparse
+import re
 import sys
+from datetime import datetime
+
 import boto3
 import botocore
-import argparse
+
+
+def list_all(client, method, type, **kwargs):
+    results = []
+    paginator = client.get_paginator(method)
+    page_iterator = paginator.paginate(**kwargs)
+    for page in page_iterator:
+        results += page[type]
+    return results
+
+
+def get_kms_key_arn(session, kms_key_id):
+    kms_client = session.client('kms')
+    pattern = re.compile(r'^arn:([^:]+):kms:([^:]*):([^:]*):((key|alias)/([^:]+))$')
+    match = pattern.match(kms_key_id)
+    if match:
+        resource_name = match.group(4)
+        if resource_name.startswith('alias'):
+            alias = list_all(kms_client, 'list_aliases', 'Aliases')
+            for alias in alias:
+                if alias['AliasName'] == resource_name:
+                    return f'arn:{match.group(1)}:kms:{match.group(2)}:{match.group(3)}:key/{alias["TargetKeyId"]}'
+        else:
+            keys = list_all(kms_client, 'list_keys', 'Keys')
+            for key in keys:
+                if key['KeyArn'] == kms_key_id:
+                    return kms_key_id
+    elif kms_key_id.startswith('alias/'):
+        alias = list_all(kms_client, 'list_aliases', 'Aliases')
+        for alias in alias:
+            if alias['AliasName'] == kms_key_id:
+                return alias['AliasArn'].replace(kms_key_id, f'key/{alias["TargetKeyId"]}')
+    else:
+        keys = list_all(kms_client, 'list_keys', 'Keys')
+        for key in keys:
+            if key['KeyId'] == kms_key_id:
+                return key['KeyArn']
+    return None
 
 
 def main(argv):
-    parser = argparse.ArgumentParser(description='Encrypts EC2 root volume.')
-    parser.add_argument('-i', '--instance',
-                        help='Instance to encrypt volume on.', required=True)
-    parser.add_argument('-key', '--customer_master_key',
-                        help='Customer master key', required=False)
-    parser.add_argument('-p', '--profile',
-                        help='Profile to use', required=False)
-    parser.add_argument('-r', '--region',
-                        help='Region of source volume', required=True)
+    parser = argparse.ArgumentParser(description='Encrypts EBS volumes of an EC2 instance.')
+    parser.add_argument('-i', '--instance', help='EC2 instance ID', required=True)
+    parser.add_argument('-k', '--kms_key_id', help='KMS key', required=True)
+    parser.add_argument('-p', '--preserve_volumes', help='Preserve original volumes',
+                        required=False, action='store_false')
     args = parser.parse_args()
 
     """ Set up AWS Session + Client + Resources + Waiters """
-    if args.profile:
-        # Create custom session
-        print('Using profile {}'.format(args.profile))
-        session = boto3.session.Session(profile_name=args.profile)
-    else:
-        # Use default session
-        session = boto3.session.Session()
+    session = boto3.session.Session()
 
-    # Get CMK
-    customer_master_key = args.customer_master_key
+    # Determine KMS key ARN
+    kms_key_arn = get_kms_key_arn(session, args.kms_key_id)
 
     client = session.client('ec2')
     ec2 = session.resource('ec2')
@@ -53,7 +82,7 @@ def main(argv):
 
     """ Check instance exists """
     instance_id = args.instance
-    print('---Checking instance ({})'.format(instance_id))
+    print('[{}] Checking instance {}'.format(datetime.now().astimezone().isoformat(), instance_id))
     instance = ec2.Instance(instance_id)
 
     try:
@@ -64,43 +93,46 @@ def main(argv):
         )
     except botocore.exceptions.WaiterError as e:
         sys.exit('ERROR: {}'.format(e))
-    
-    all_mappings = []    
-    
+
+    all_mappings = []
+
     block_device_mappings = instance.block_device_mappings
 
     for device_mapping in block_device_mappings:
         original_mappings = {
             'DeleteOnTermination': device_mapping['Ebs']['DeleteOnTermination'],
-            'VolumeId': device_mapping['Ebs']['VolumeId'],
             'DeviceName': device_mapping['DeviceName'],
+            'VolumeId': device_mapping['Ebs']['VolumeId'],
         }
         all_mappings.append(original_mappings)
-  
+
     volume_data = []
-    
-    print('---Preparing instance')    
-    """ Get volume and exit if already encrypted """
+
+    """ Get volume and exit if already encrypted with the provided KMS key """
+
+    has_volumes_to_encrypt = False
     volumes = [v for v in instance.volumes.all()]
     for volume in volumes:
         volume_encrypted = volume.encrypted
-        
+
         current_volume_data = {}
         for mapping in all_mappings:
             if mapping['VolumeId'] == volume.volume_id:
                 current_volume_data = {
-                    'volume': volume,
                     'DeleteOnTermination': mapping['DeleteOnTermination'],
                     'DeviceName': mapping['DeviceName'],
-                }        
-                 
-        if volume_encrypted:
-            sys.exit(
-                '**Volume ({}) is already encrypted'
-                .format(volume.id))
+                    'volume': volume,
+                }
+
+        if volume_encrypted and volume.kms_key_id == kms_key_arn:
+            print('[{}] Volume {} is already encrypted with KMS key {}'.format(
+                datetime.now().astimezone().isoformat(), volume.id, kms_key_arn))
+            continue
+        else:
+            has_volumes_to_encrypt = True
 
         """ Step 1: Prepare instance """
-    
+
         # Exit if instance is pending, shutting-down, or terminated
         instance_exit_states = [0, 32, 48]
         if instance.state['Code'] in instance_exit_states:
@@ -108,14 +140,15 @@ def main(argv):
                 'ERROR: Instance is {} please make sure this instance is active.'
                 .format(instance.state['Name'])
             )
-    
+
         # Validate successful shutdown if it is running or stopping
-        if instance.state['Code'] is 16:
+        if instance.state['Code'] == 16:
+            print('[{}] Stopping instance {}'.format(datetime.now().astimezone().isoformat(), instance_id))
             instance.stop()
-    
+
         # Set the max_attempts for this waiter (default 40)
         waiter_instance_stopped.config.max_attempts = 80
-    
+
         try:
             waiter_instance_stopped.wait(
                 InstanceIds=[
@@ -124,16 +157,17 @@ def main(argv):
             )
         except botocore.exceptions.WaiterError as e:
             sys.exit('ERROR: {}'.format(e))
-    
+
         """ Step 2: Take snapshot of volume """
-        print('---Create snapshot of volume ({})'.format(volume.id))
+        print('[{}] Create snapshot of volume {} ({})'.format(
+            datetime.now().astimezone().isoformat(), volume.id, current_volume_data['DeviceName']))
         snapshot = ec2.create_snapshot(
+            Description='Snapshot of volume {}'.format(volume.id),
             VolumeId=volume.id,
-            Description='Snapshot of volume ({})'.format(volume.id),
         )
-        
+
         waiter_snapshot_complete.config.max_attempts = 240
-    
+
         try:
             waiter_snapshot_complete.wait(
                 SnapshotIds=[
@@ -143,69 +177,65 @@ def main(argv):
         except botocore.exceptions.WaiterError as e:
             snapshot.delete()
             sys.exit('ERROR: {}'.format(e))
-    
+
         """ Step 3: Create encrypted volume """
-        print('---Create encrypted copy of snapshot')
-        if customer_master_key:
-            # Use custom key
-            snapshot_encrypted_dict = snapshot.copy(
-                SourceRegion=args.region,
-                Description='Encrypted copy of snapshot #{}'
-                            .format(snapshot.id),
-                KmsKeyId=customer_master_key,
-                Encrypted=True,
-            )
-        else:
-            # Use default key
-            snapshot_encrypted_dict = snapshot.copy(
-                SourceRegion=args.region,
-                Description='Encrypted copy of snapshot ({})'
-                            .format(snapshot.id),
-                Encrypted=True,
-            )
-    
-        snapshot_encrypted = ec2.Snapshot(snapshot_encrypted_dict['SnapshotId'])
-    
-        try:
-            waiter_snapshot_complete.wait(
-                SnapshotIds=[
-                    snapshot_encrypted.id,
-                ],
-            )
-        except botocore.exceptions.WaiterError as e:
-            snapshot.delete()
-            snapshot_encrypted.delete()
-            sys.exit('ERROR: {}'.format(e))
-    
-        print('---Create encrypted volume from snapshot')
+        print('[{}] Create encrypted volume from snapshot {}'.format(
+            datetime.now().astimezone().isoformat(), snapshot.id))
 
         if volume.volume_type == 'io1':
             volume_encrypted = ec2.create_volume(
-                SnapshotId=snapshot_encrypted.id,
-                VolumeType=volume.volume_type,
+                AvailabilityZone=instance.placement['AvailabilityZone'],
+                Encrypted=True,
                 Iops=volume.iops,
-                AvailabilityZone=instance.placement['AvailabilityZone']
+                KmsKeyId=kms_key_arn,
+                SnapshotId=snapshot.id,
+                VolumeType=volume.volume_type,
             )
         else:
             volume_encrypted = ec2.create_volume(
-                SnapshotId=snapshot_encrypted.id,
+                AvailabilityZone=instance.placement['AvailabilityZone'],
+                Encrypted=True,
+                KmsKeyId=kms_key_arn,
+                SnapshotId=snapshot.id,
                 VolumeType=volume.volume_type,
-                AvailabilityZone=instance.placement['AvailabilityZone']
             )
-     
-        # Add original tags to new volume 
+
+        # Add original tags to new volume
         if volume.tags:
-            volume_encrypted.create_tags(Tags=volume.tags)
-    
+            volume_encrypted.create_tags(Tags=[t for t in volume.tags if not t.get(
+                'Key').startswith('VolumeEncryptionMetadata:')])
+
+        # Add additional metadata tags to original volumes for traceability
+        metadata_tags = [
+            {
+                'Key': 'VolumeEncryptionMetadata:DeviceName',
+                'Value': current_volume_data['DeviceName']
+            },
+            {
+                'Key': 'VolumeEncryptionMetadata:InstanceId',
+                'Value': instance_id
+            }
+        ]
+        for t in instance.tags:
+            if t['Key'] == 'Name':
+                metadata_tags.append({
+                    'Key': 'VolumeEncryptionMetadata:InstanceName',
+                    'Value': t['Value']
+                })
+                break
+        volume.create_tags(Tags=metadata_tags)
+
         """ Step 4: Detach current volume """
-        print('---Detach volume {}'.format(volume.id))
+        print('[{}] Detach volume {} ({})'.format(datetime.now().astimezone(
+        ).isoformat(), volume.id, current_volume_data['DeviceName']))
         instance.detach_volume(
+            Device=current_volume_data['DeviceName'],
             VolumeId=volume.id,
-            Device=current_volume_data['DeviceName']
         )
-    
+
         """ Step 5: Attach new encrypted volume """
-        print('---Attach volume {}'.format(volume_encrypted.id))
+        print('[{}] Attach volume {} ({})'.format(datetime.now().astimezone().isoformat(),
+              volume_encrypted.id, current_volume_data['DeviceName']))
         try:
             waiter_volume_available.wait(
                 VolumeIds=[
@@ -214,19 +244,17 @@ def main(argv):
             )
         except botocore.exceptions.WaiterError as e:
             snapshot.delete()
-            snapshot_encrypted.delete()
             volume_encrypted.delete()
             sys.exit('ERROR: {}'.format(e))
-    
+
         instance.attach_volume(
+            Device=current_volume_data['DeviceName'],
             VolumeId=volume_encrypted.id,
-            Device=current_volume_data['DeviceName']
         )
-        
+
         current_volume_data['snapshot'] = snapshot
-        current_volume_data['snapshot_encrypted'] = snapshot_encrypted
-        volume_data.append(current_volume_data)                  
-    
+        volume_data.append(current_volume_data)
+
     for bdm in volume_data:
         # Modify instance attributes
         instance.modify_attribute(
@@ -240,29 +268,35 @@ def main(argv):
                 },
             ],
         )
-    """ Step 6: Start instance """
-    print('---Start instance')
-    instance.start()
-    try:
-        waiter_instance_running.wait(
-            InstanceIds=[
-                instance_id,
-            ]
-        )
-    except botocore.exceptions.WaiterError as e:
-        sys.exit('ERROR: {}'.format(e))
-    
-    """ Step 7: Clean up """
-    print('---Clean up resources')
-    for cleanup in volume_data:
-        print('---Remove snapshot {}'.format(cleanup['snapshot'].id))
-        cleanup['snapshot'].delete()
-        print('---Remove encrypted snapshot {}'.format(cleanup['snapshot_encrypted'].id))
-        cleanup['snapshot_encrypted'].delete()
-        print('---Remove original volume {}'.format(cleanup['volume'].id))
-        cleanup['volume'].delete()
-    
-    print('Encryption finished')
+
+    if has_volumes_to_encrypt:
+        """ Step 6: Start instance """
+        print('[{}] Start instance {}'.format(datetime.now().astimezone().isoformat(), instance.id))
+        instance.start()
+        try:
+            waiter_instance_running.wait(
+                InstanceIds=[
+                    instance_id,
+                ]
+            )
+        except botocore.exceptions.WaiterError as e:
+            sys.exit('ERROR: {}'.format(e))
+
+        """ Step 7: Clean up """
+        print('[{}] Clean up resources'.format(datetime.now().astimezone().isoformat()))
+        for cleanup in volume_data:
+            print('[{}] Delete snapshot {}'.format(datetime.now().astimezone().isoformat(), cleanup['snapshot'].id))
+            cleanup['snapshot'].delete()
+            if not args.preserve_volumes:
+                print('[{}] Skipping deletion of original volume {} ({})'.format(
+                    datetime.now().astimezone().isoformat(), cleanup['volume'].id, cleanup['DeviceName']))
+            else:
+                print('[{}] Delete original volume {} ({})'.format(
+                    datetime.now().astimezone().isoformat(), cleanup['volume'].id, cleanup['DeviceName']))
+                cleanup['volume'].delete()
+
+    print('[{}] Encryption finished'.format(datetime.now().astimezone().isoformat()))
+
 
 if __name__ == "__main__":
     main(sys.argv[1:])
